Time: 159 ms (64.59%), Space: 24.4 MB (59.33%) - LeetHub

+++ 2222-number-of-ways-to-select-buildings/2222-number-of-ways-to-select-buildings.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long numberOfWays(string s) {
         int N = s.size();
-        long number_of_ways = 0;
+        long long number_of_ways = 0;
         vector<long long> middle(2), first_order(2); 
         for(int i = 0; i < N; i++){
             number_of_ways += middle['1' - s[i]];
